refactor(reborn): tighten routine preview projection
